# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CliniData Backend API
  version: 1.0.0
  description: API mínima (MVP) para CliniData — autenticación, usuarios, pacientes, consultas y auditoría. Seguridad por JWT Bearer.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/users/me:
    get:
      tags: [users]
      summary: Obtener usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { description: Unauthorized }
  /api/users:
    get:
      tags: [users]
      summary: Listar médicos (ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403': { description: Forbidden }
  /api/users/{id}/status:
    put:
      tags: [users]
      summary: Activar/desactivar médico (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                activo:
                  type: boolean
      responses:
        '200': { description: OK }
        '400': { description: Bad request }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        nombre: { type: string }
        apellido: { type: string }
        rol:
          type: string
          enum: [ADMIN, MEDICO, USER]
        activo: { type: boolean }
        clinicaId: { type: string }
...